+++ ios/Classes/FlutterNativeTimezonePlugin.h
@@ -1,0 +1,5 @@
+#import <CoreLocation/CoreLocation.h>
+#import <Flutter/Flutter.h>
+
+@interface FlutterNativeTimezonePlugin : NSObject<FlutterPlugin>
+@end