+++ test/flutter_native_timezone_test.dart
@@ -1,0 +1,21 @@
+import 'package:flutter/services.dart';
+import 'package:flutter_native_timezone/flutter_native_timezone.dart';
+import 'package:flutter_test/flutter_test.dart';
+
+void main() {
+  const MethodChannel channel = MethodChannel('flutter_native_timezone');
+
+  setUp(() {
+    channel.setMockMethodCallHandler((MethodCall methodCall) async {
+      return '42';
+    });
+  });
+
+  tearDown(() {
+    channel.setMockMethodCallHandler(null);
+  });
+
+  test('getPlatformVersion', () async {
+    expect(await FlutterNativeTimezone.getLocalTimezone(), '42');
+  });
+}